Added total number of participants and info of how many are missing bioklubbkortsnummer
